Clarify object model and hook stuff up
Renaming:

Allocation -> Unit
allocate -> buy

I didn't want to store the counts/totals in the object classes so
I put them into a player object. It makes the knockout a little more
clunky but it's work in progress and that can always be put into
accessors on the view model.